Extract SetupGame class
